IN-224 Added posibility to specify time slots for schedule-holiday mappings.

+++ src/access-group-schedule-zone/access-group-schedule-zone.service.ts
@@ -11,6 +11,10 @@ export class AccessGroupScheduleZoneService {
   ) {}
 
   async removeAll(accessGroupScheduleZones: AccessGroupScheduleZone[]) {
+    if (!accessGroupScheduleZones || !accessGroupScheduleZones.length) {
+      return [];
+    }
+
     return await this.accessGroupScheduleZoneRepository.remove(
       accessGroupScheduleZones,
     );
